Refactor styling
