+++ client/coral-admin/src/routes/Community/components/ActionButton.css
@@ -1,0 +1,4 @@
+.actionButton {
+  transform: scale(.8);
+  margin: 0;
+}
