+++ client/coral-admin/src/routes/Community/components/styles.css
@@ -1,118 +1,50 @@
 @custom-media --big-viewport (min-width: 780px);
 
-.listContainer {
-  max-width: 860px;
-  margin: 0 auto;
-}
-
-.tabBar {
-  background-color: rgba(44, 44, 44, 0.89);
-  z-index: 5;
-}
-
-.tab {
-  flex: 1;
-  color: white;
-  text-transform: capitalize;
-  font-weight: 500;
-  font-size: 15px;
-  letter-spacing: 1px;
-  transition: border-bottom 200ms;
-}
-
-.active {
-  color: white;
-  box-sizing: border-box;
-  border-bottom: solid 5px #F36451;
-}
-
-.active > span {
-  color: white;
-}
-
-.active:after {
-  background: transparent !important;
-}
-
-.showShortcuts {
-  position: absolute;
-  right: 130px;
+.container {
+  padding: 10px;
   display: flex;
-  align-items: center;
-  font-size: 13px;
-
-span {
-  margin-left: 7px;
-}
+  padding-bottom: 200px;
 }
 
-@media (--big-viewport) {
-  .tab {
-    flex: none;
-  }
+.leftColumn {
+  padding: 42px 56px;
+  width: 234px;
 }
 
-.approve {
-  margin-top: 10px;
+.mainContent {
+  width: calc(100% - 300px);
+  padding: 34px 14px;
+  box-sizing: border-box;
 }
 
-.notFound {
-  position: relative;
-  margin: 20px auto;
-  text-align: center;
-  padding: 68px 45px;
-  vertical-align: middle;
-  min-width: 500px;
-
-  a {
-    color: rgb(244, 126, 107);
-    font-weight: 500;
-
-    &.goToStreams {
-      position: absolute;
-      right: 10px;
-      bottom: 10px;
-    }
-  }
+.roleButton {
+  display: block;
 }
 
-.header {
-  background-color: #2c2c2c;
-  color: white;
-  margin-bottom: -1px;
+.searchBox {
+  width: 100%;
+  padding: 9px;
+  border: 1px solid #ccc;
+  border-radius: 2px;
+  display: flex;
+  background: white;
+  box-sizing: border-box;
+  height: 40px;
 
-  .settingsButton {
-    i {
-      vertical-align: middle;
-      margin-left: 10px;
-      margin-top: -4px;
-    }
+  i {
+    color: #A1A1A1
   }
 
-  .moderateAsset {
-    a {
-      text-align: center;
-      -webkit-box-flex: 1;
-      -ms-flex: 1;
-      flex: 1;
-      color: white;
-      text-transform: capitalize;
-      font-weight: 500;
-      font-size: 15px;
-      letter-spacing: 1px;
-      transition: opacity 200ms;
-      opacity: 1;
-
-      &:hover {
-         opacity: .8;
-         cursor: pointer;
-      }
-    }
+  input {
+    display: block;
+    width: 100%;
+    height: 100%;
+    border: none;
+    font-size: 16px;
+    padding: 0 2px 0 15px;
+    box-sizing: border-box;
   }
 }
-
-
-@custom-media --big-viewport (min-width: 780px);
 
 .list {
   padding: 8px 0;
@@ -134,14 +66,11 @@ span {
       position: fixed;
       bottom: 60px;
       left: 25%;
+      margin: 0 auto;
       display: flex;
       justify-content: space-around;
       width: 50%;
       margin: 0;
-    }
-
-    .actionButton {
-      transform: scale(1.4);
     }
   }
 }
@@ -159,6 +88,7 @@ span {
   min-height: 220px;
 
   .container {
+    display: block;
     padding: 0 14px;
     min-height: 180px;
   }
@@ -184,9 +114,9 @@ span {
      right: 0;
      height: 100%;
      top: 0;
-     padding: 40px 18px;
+     padding: 50px 18px;
      box-sizing: border-box;
-  }
+   }
 
   .itemHeader {
     display: flex;
@@ -194,6 +124,8 @@ span {
     justify-content: space-between;
 
     .author {
+      font-size: 16px;
+      font-weight: bold;
       min-width: 230px;
       display: flex;
       align-items: center;
@@ -221,18 +153,12 @@ span {
     margin-left: 40px;
   }
 
-  .actionButton {
-    transform: scale(.8);
-    margin: 0;
-  }
-
   .body {
     margin-top: 0px;
     flex: 1;
     color: black;
     max-width: 500px;
     word-wrap: break-word;
-    font-weight: 300;
   }
 
   .flagged {
@@ -245,10 +171,7 @@ span {
     font-size: 12px;
     color: #d32f2f;
   }
-
 }
-
-
 
 .empty {
   color: #444;
@@ -301,34 +224,55 @@ span {
   margin-top: 5px;
 }
 
-.Comment {
-  .moderateArticle {
-    font-size: 12px;
-    a {
-      display: inline-block;
-      color: #679af3;
-      text-decoration: none;
-      font-size: 1em;
-      font-weight: 400;
-      letter-spacing: .5px;
-      font-size: 12px;
-      margin-left: 10px;
+.banButton {
+  width: 114px;
+  letter-spacing: 1px;
 
-      &:hover {
-        text-decoration: underline;
-        opacity: .9;
-        cursor: pointer;
-       }
-    }
+  i {
+    vertical-align: middle;
+    margin-right: 10px;
+    font-size: 14px;
   }
 }
 
-.flagBox {
-  max-width: 480px;
-  border-top: 1px solid rgba(66, 66, 66, 0.12);
-  h3 {
-    font-size: 14px;
-    margin: 0;
-    font-weight: 500;
+.flaggedByCount {
+  display: block;
+  text-align: left;
+  margin-top: 5px;
+}
+
+.flaggedByCount i {
+  font-size: 14px;
+  margin-right: 10px;
+}
+
+.flaggedBy {
+  display: inline;
+  padding: 3px;
+  font-size: 14px;
+  margin-left: 5px;
+}
+
+.flaggedByLabel {
+  font-weight: 600;
+  font-size: 14px;
+}
+
+.flaggedReasons {
+  margin-left: 24px;
+  margin-top: 10px;
+}
+
+p.flaggedByReason {
+  font-size: 1tpx;
+  margin: 0px;
+  line-height: 1.4;
+}
+
+.button {
+  composes: buttonReset from 'coral-framework/styles/reset.css';
+  vertical-align: text-bottom;
+  &:hover {
+    background-color: #E0E0E0;
   }
 }
